pyproject.toml
[project]
name = "real-estate-location-selection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4==4.13.4",
    "browserforge==1.2.3",
    "bs4==0.0.2",
    "camoufox==0.4.11",
    "falcon==4.0.2",
    "google-cloud-batch==0.17.26",
    "google-cloud-bigquery==3.25.0",
    "google-cloud-bigquery-storage==2.25.0",
    "google-cloud-compute>=1.32.0",
    "google-cloud-pubsub==2.29.0",
    "google-cloud-secret-manager==2.23.3",
    "google-cloud-storage==3.1.0",
    "gunicorn==23.0.0",
    "haversine==2.9.0",
    "json5==0.12.0",
    "lxml==5.4.0",
    "nest-asyncio>=1.6.0",
    "psycopg==3.2.9",
    "psycopg-binary==3.2.9",
    "pyarrow==20.0.0",
    "pytest-playwright==0.7.0",
    "requests==2.32.3",
    "tabulate>=0.9.0",
]
